requirements.txt
flask==3.1.0
flask-wtf==1.2.2
flask-login==0.6.3
pymongo==4.11.2
python-dotenv==1.0.1
email-validator==2.2.0
flask-bcrypt==1.0.1
gunicorn==21.2.0
Pillow==11.1.0 